Extract getTypeName and getNamespaceString outside of checker
