+++ packages/compiler/core/decorator-utils.ts
@@ -1,5 +1,5 @@
 import { getIntrinsicModelName, getPropertyType } from "../lib/decorators.js";
-import { DecoratorContext } from "./index.js";
+import { DecoratorContext, getTypeName } from "./index.js";
 import { createDiagnostic, reportDiagnostic } from "./messages.js";
 import { Program } from "./program.js";
 import {
@@ -322,7 +322,7 @@ export function validateDecoratorParamCount(
 
 function prettyValue(program: Program, value: any) {
   if (typeof value === "object" && value !== null && "kind" in value) {
-    return program.checker.getTypeName(value);
+    return getTypeName(value);
   }
   return value;
 }
