+++ packages/compiler/test/semantic-walker.test.ts
@@ -1,5 +1,6 @@
 import { deepStrictEqual, ok, strictEqual } from "assert";
 import {
+  getNamespaceString,
   Interface,
   ListenerFlow,
   Model,
@@ -120,7 +121,7 @@ describe("compiler: semantic walker", () => {
     `);
 
     deepStrictEqual(
-      result.namespaces.map((x) => host.program.checker.getNamespaceString(x)),
+      result.namespaces.map((x) => getNamespaceString(x)),
       ["", "Global", "Global.My", "Global.My.Simple", "Global.My.Parent", "Global.My.Parent.Child"]
     );
   });
