+++ packages/openapi/src/helpers.ts
@@ -1,6 +1,7 @@
 import {
   getFriendlyName,
   getServiceNamespace,
+  getTypeName,
   getVisibility,
   isGlobalNamespace,
   isTemplateInstance,
@@ -49,13 +50,13 @@ export function shouldInline(program: Program, type: Type): boolean {
  * Cadl-native names are shortened to exclude root `Cadl` namespace and service
  * namespace using the provided `TypeNameOptions`.
  */
-export function getTypeName(
+export function getOpenAPITypeName(
   program: Program,
   type: Type,
   options: TypeNameOptions,
   existing?: Record<string, any>
 ): string {
-  const name = getFriendlyName(program, type) ?? program.checker.getTypeName(type, options);
+  const name = getFriendlyName(program, type) ?? getTypeName(type, options);
 
   checkDuplicateTypeName(program, type, name, existing);
   return name;
@@ -89,7 +90,7 @@ export function getParameterKey(
   options: TypeNameOptions
 ): string {
   const parent = property.model!;
-  let key = getTypeName(program, parent, options);
+  let key = getOpenAPITypeName(program, parent, options);
 
   if (parent.properties.size > 1) {
     key += `.${property.name}`;
